requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.27.0
requests==2.31.0
xmltodict==0.13.0
